fix: give the catalog card to a browse that has a catalog
The rule picked the last browse in the message by position alone, so a
second search that was still running — or that came back with an error
or an empty server list — took the card from an earlier one that had a
readable catalog. On screen the catalog appeared, blanked when the model
searched again, and came back only if that search returned something
renderable.

A browse now has to carry a payload the card can render before it can
hold it, and the card asks that same question of its own result rather
than checking status separately, so the two cannot disagree. The parser
moves next to the browse predicate to make that possible. A lone browse
still waiting on its result keeps the generic card, which is where its
running state shows; it is only suppressed while another browse is
actually drawing.
